refactor: 컴포넌트 단위로 뷰 및 테스트 코드 단순 분리
