+++ README.md
@@ -1,10 +1,26 @@
 # android-signup
 
-## Step2 기능 목록
+## Step4 구현 목록
 
-- [x] 타이틀 추가
-- [x] 입력 폼 추가
-- [x] 회원가입 버튼 추가
+- [x] 컴포넌트 단위로 뷰 및 테스트 코드 단순 분리
+- [] SignUpTitle 테스트 코드 작성
+- [] SignUpHelperText 테스트 코드 작성
+- [] 컴포넌트 공통 요소 추출
+- [] 컴포넌트 고유 속성 주입
+- [] 테스트 코드 리팩터링
+- [] 유효성 검사 로직 테스트 코드 추가
+- [] SignUpButton 활성화 로직 추가
+- [] SignUpButton 활성화 뷰 반영
+- [] SignUpButton 테스트 코드 작성
+-
+
+## Step4 진행 중 의식의 흐름
+
+- 리팩터링을 하면서 테스트 코드가 컴포넌트 단위로 나눠지지 않은 것이 문제였다는 생각이 듭니다. 수정한 부분 외에 모든 컴포넌트 테스트가 같이 돌아가니 너무 느리네요. 이건
+  컴포즈의 문제가 아니라 저의 명백한 실수네요.
+- 하나의 TextField로 모든 컴포넌트를 커버하기 위해서 불필요한 로직(복잡성!) 고민으로 보낸 시간이 아깝네요. 그 시간에 컴포넌트 단위로 나누는 연습에 더 집중했어야
+  했는데...
+-
 
 ## Step3 구현 목록
 
@@ -39,3 +55,9 @@
   InitViews에서 별도로 HelperText를 관리하지 않아도 될 것 같다는 생각이 들었습니다.
   리팩터링 해볼 만한 요소일까요? 무리하게 하나의 컴포저블 함수로 관리하려다 보니 함수 내부가 비대해 지는 것을 느꼈습니다. 개선에 대한 고민이 필요할 것 같습니다.
 - 테스트 코드를 짜면서 뭔가 단단히 잘못됨을 느꼈다. 특히 비밀번호 일치 테스트 코드는 엉망인 것 같다. 이 부분을 개선해야겠다.
+
+## Step2 기능 목록
+
+- [x] 타이틀 추가
+- [x] 입력 폼 추가
+- [x] 회원가입 버튼 추가
