+++ app/src/main/java/nextstep/signup/view/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.theme
+package nextstep.signup.view.ui.theme
 
 import androidx.compose.ui.graphics.Color
