+++ app/src/main/java/nextstep/signup/view/ui/theme/Dimens.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.theme
+package nextstep.signup.view.ui.theme
 
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
