+++ app/src/main/java/nextstep/signup/view/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.theme
+package nextstep.signup.view.ui.theme
 
 import android.app.Activity
 import android.os.Build
