+++ app/src/main/java/nextstep/signup/view/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.theme
+package nextstep.signup.view.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
